backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
psycopg2-binary
pydantic
face-recognition
numpy
Pillow
